testing-container-rd/requirements.txt
Flask==0.10.1
kafka-python
elasticsearch
pymongo
requests